Adjust for generating an `oemof` sub-package
The most important thing is changing `setup`'s `packages` parameter in
"setup.py" so that it always installs an `"oemof"` package along the other
ones it does find and to make sure that it searches in subdirectories of
the `"oemof"` directory, because it doesn't go into that one by itself as
that one doesn't contain an "__init__.py" file.
Also `nose` doesn't seem to like test file names with dots in them so
that one got moved. The "oemof.network.py" got moved for obvious
reasons.