Make `Node.registry` a class property
This is necessary in order to intercept read and write access to the
attribute in order to warn users about the deprecation of this feature.
Since this can only be done by defining the property in `Node`'s
metaclass, I had to explicitly create one in order to do so.