Make `Node` creation only emit necessary warnings
When creating a `Node`, warnings should only emitted if the `Node` is
then added to an `EnergySystem` via the automatic registration
mechanism. This commit switches of warnings temporarily in order to
check whether `Node.registry` is `None`. If that is the case, nothing
happens, so no warning should be emitted. If it is an `EnergySystem`,
the `Node` will be added to it, which should trigger the desired
warning.