Make our style checkers happy
The imports had to be reordered because `isort` wants "`from`" imports
after the others. The extraneous empty line also had to be removed
because of `isort` complaints.
The new test modules docstring's first line was to long for the style
checker, so I used `.format` to keep the line length below 140
characters. I couldn't break the line because the first line of a
docstring is special.
The pandas import had to be removed because it wasn't used and the `n =`
lines had to be changed because the assigned variable wasn't used.