Use `Node.register` instead of its inlined code
I have no idea why I didn't just call the corresponding method in the
first place. But now that the method makes sure that no inappropriate
warnings are emitted, it's actually important to centralize this in
order for it to be handled correctly.
All tests still pass, so refactoring this should be OK.