Implement recursively `add`ing a subnodes
This is a two step process. First, on construction of a node that
possibly has subnodes, the node in question connects to a blinker signal
that notifies the node when it gets added to an EnergySystem. When this
happens, the `add_subnodes` callback is called, which adds the subnodes
the EnergySystem, too.
Note that in order to `connect` to the blinker signal, `EnergySystem`
has to be `import`ed, which needs to be done via a local import, in
order to avoid a circular import.