Remove the `interface` switch from `SubNetworkLabel`
Remove the switch in order to find a better place for it. The purpose of
the switch was to keep track of whether a node is and "interface node",
i.e. whether it connects to a node outside of its `SubNetwork` or not.
A possible problem is, that the `SubNetworkLabel` has to be constructed
at the same time that the subnode is constructed and at this time, it
might not be known whether the node is and interface node or not.
IMHO a better place for this switch would be as a property on the
subnode, which could automatically infer this value by looking at the
inputs and outputs.