Define a depth property for Node
Add a `depth` property to Node and set a default for it to 1 if the
label is not a HierachicalLabel instance. It is directly useful for
oemof-visio, for example to build the graph in presence of SubNetwork
instances, one can list the relevant nodes for a given depth simply with
[n for n in energy_system.nodes if n.depth == depth and not
isinstance(n, SubNetwork)]. It might also proove itself useful for
result processing. The name `depth` is just a suggestion and can be
changed to something else.