Remove HierachicalLabel class
The class Node now has a "parent" attribute by itself.
If `SubNetwork.subnode` is not used, providing unique labels
is up to the user. The named function accepts a `local_name`
instead of the label and creates a unique tuple using the unique label
of the `SubNetwork` and the (locally unique) local name.