Create branch to unify AtomicNode and SubNetwork
The idea behind this branch is to allow every Node to have a
sub-structure (code name "fractal nodes"). From a mathematical
point of view, even simple linear components could be expressed
by a collection of sub-components. If we eliminate the necessity
to define something as atomic, this would be reflected in our mode
and give more flexibility.